+++ libs/providers/langchain-mcp-adapters/package.json
@@ -15,16 +15,16 @@
     "url": "https://github.com/langchain-ai/langchainjs/issues"
   },
   "scripts": {
-    "build": "run-s build:main",
+    "build": "run-s build:main build:examples",
     "build:main": "pnpm --filter @langchain/build compile @langchain/mcp-adapters",
+    "build:examples": "tsc -p ./examples/tsconfig.json",
     "clean": "rm -rf dist/ dist-cjs/ .turbo/",
     "format": "prettier --config .prettierrc --write \"src/**/*.ts\" \"examples/**/*.ts\"",
     "format:check": "prettier --config .prettierrc --check \"src\" \"examples/**/*.ts\"",
-    "lint": "run-s lint:eslint lint:dpdm lint:examples",
+    "lint": "run-s lint:eslint lint:dpdm",
     "lint:dpdm": "dpdm --skip-dynamic-imports circular --exit-code circular:1 --no-warning --no-tree src/**/*.ts examples/**/*.ts",
     "lint:eslint": "eslint --cache src/ examples/",
     "lint:fix": "pnpm lint:eslint --fix && pnpm lint:dpdm",
-    "lint:examples": "tsc -p ./examples/tsconfig.json",
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
     "test:watch": "vitest"
